global refactor from eager to lazy

+++ src/main/java/com/cusbee/yoki/serviceimpl/DishServiceImpl.java
@@ -1,5 +1,6 @@
 package com.cusbee.yoki.serviceimpl;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 import java.util.regex.Matcher;
@@ -15,9 +16,11 @@ import com.cusbee.yoki.entity.Category;
 import com.cusbee.yoki.entity.CrudOperation;
 import com.cusbee.yoki.entity.Dish;
 import com.cusbee.yoki.entity.DishType;
+import com.cusbee.yoki.entity.Ingredient;
 import com.cusbee.yoki.exception.ApplicationException;
 import com.cusbee.yoki.exception.BaseException;
 import com.cusbee.yoki.model.DishModel;
+import com.cusbee.yoki.model.IngredientModel;
 import com.cusbee.yoki.repositories.DishRepository;
 import com.cusbee.yoki.service.DishService;
 import com.cusbee.yoki.service.IngredientService;
@@ -54,13 +57,29 @@ public class DishServiceImpl implements DishService {
 	}
 
 	@Override
-	public void remove(Dish dish) {
+	@Transactional
+	public void remove(Long id) throws BaseException {
+		Dish dish = get(id);
+		if(Objects.isNull(dish)){
+			throw new ApplicationException(ErrorCodes.Dish.EMPTY_REQUEST, "This dish are not present");
+		}
+		dish.setCategory(null);
+		List<Ingredient> ingredients = dish.getIngredients();
+		dish.getIngredients().removeAll(ingredients);
+		for(Ingredient ingredient : ingredients) {
+			ingredient.getDish().remove(dish);
+		}
 		this.dao.remove(dish);
 	}
 
 	@Override
-	public Dish getById(Long id) {
-		return this.dao.getById(id);
+	@Transactional
+	public Dish get(Long id) throws BaseException {
+		Dish dish = this.dao.get(id);
+		if(Objects.isNull(dish)){
+			throw new ApplicationException(ErrorCodes.Dish.EMPTY_REQUEST, "Dish with this ID are not present");
+		}
+		return dish;
 	}
 
 	@Override
@@ -109,14 +128,14 @@ public class DishServiceImpl implements DishService {
 				dish.setType(DishType.valueOf(request.getType().toUpperCase()));
 			}
 			if(!Objects.isNull(request.getCategoryId())){
-				Category category = categoryDao.getById(request.getCategoryId());
+				Category category = categoryDao.get(request.getCategoryId());
 				nullPointerService.isNull(category);
 				dish.setCategory(category);
 				category.getDishes().add(dish);
 			}
 			return dish;
 		case UPDATE:
-			dish = getById(request.getId());
+			dish = get(request.getId());
 			nullPointerService.isNull(dish);
 			if(!validate(request.getName())){
 				throw new ApplicationException(ErrorCodes.Dish.INVALID_REQUEST, "Name: " + request.getName() +" already exists");
@@ -140,7 +159,7 @@ public class DishServiceImpl implements DishService {
 				dish.setType(DishType.valueOf(request.getType().toUpperCase()));
 			}
 			if(!Objects.isNull(request.getCategoryId())){
-				Category category = categoryDao.getById(request.getCategoryId());
+				Category category = categoryDao.get(request.getCategoryId());
 				nullPointerService.isNull(category);
 				category.getDishes().remove(dish);
 				dish.setCategory(category);
@@ -151,6 +170,23 @@ public class DishServiceImpl implements DishService {
 			throw new ApplicationException(ErrorCodes.Common.INVALID_REQUEST,
 					"Invalid request");
 		}
+	}
+	
+	public Dish addIngredients(DishModel request) throws BaseException {
+		if(Objects.isNull(request.getId())){
+			throw new ApplicationException(ErrorCodes.Dish.EMPTY_FIELD, "Field ID is not present");
+		}
+		if(Objects.isNull(request.getIngredients())){
+			throw new ApplicationException(ErrorCodes.Ingredient.EMPTY_REQUEST, "Ingredients are empty");
+		}
+		Dish dish = get(request.getId());
+		List<Ingredient> ingredients = new ArrayList<>();
+		for(IngredientModel ingredient: request.getIngredients()){
+			Ingredient ingred = ingredientService.get(ingredient.getId());
+			ingredients.add(ingred);
+		}
+		dish.getIngredients().addAll(ingredients);
+		return dish;
 	}
 	
 	protected boolean validateName(String name) throws BaseException {
